Move WorkflowFunction to CmsBundle
