+++ src/OpenOrchestra/Tests/Functional/ModelBundle/Repository/WorkflowFunctionRepositoryTest.php
@@ -5,7 +5,7 @@ namespace OpenOrchestra\FunctionalTests\ModelBundle\Repository;
 use OpenOrchestra\BaseBundle\Tests\AbstractTest\AbstractKernelTestCase;
 use OpenOrchestra\Pagination\Configuration\FinderConfiguration;
 use OpenOrchestra\Pagination\Configuration\PaginateFinderConfiguration;
-use OpenOrchestra\WorkflowFunction\Repository\WorkflowFunctionRepositoryInterface;
+use OpenOrchestra\Workflow\Repository\WorkflowFunctionRepositoryInterface;
 
 /**
  * Class WorkflowFunctionRepositoryTest
